fix editable text functionality

+++ src/app/directives/keyboard-commands.directive.ts
@@ -15,20 +15,36 @@ export class KeyboardCommandsDirective {
 
   @HostListener('keydown.control.v', ['$event'])
   @HostListener('keydown.meta.v', ['$event'])
-  onPaste() {
+  onPaste($event: KeyboardEvent) {
+    if (this.isTextElementInEditMode($event.target)) {
+      return;
+    }
+
     this.copyPasteService.paste();
   }
 
   @HostListener('keydown.control.c', ['$event'])
   @HostListener('keydown.meta.c', ['$event'])
   onCopy($event: KeyboardEvent) {
+    if (this.isTextElementInEditMode($event.target)) {
+      return;
+    }
+
     $event.stopPropagation();
     this.copyPasteService.copy();
   }
 
   @HostListener('keydown.delete', ['$event'])
   onDelete($event: KeyboardEvent) {
+    if (this.isTextElementInEditMode($event.target)) {
+      return;
+    }
+
     $event.stopPropagation();
     this.canvasService.deleteItem(this.selectionService.selectedItem?.key);
+  }
+
+  private isTextElementInEditMode(eventTarget: EventTarget | null): boolean {
+    return (eventTarget as HTMLElement).contentEditable === 'true';
   }
 }
